pyproject.toml
[project]
name = "repo-fittude"
version = "0.1.0"
description = "Repo functions and querys to fittude application"
authors = [
    {name = "Luiz Carlos Souza Garcia",email = "lcsouzagarcia@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.12"

